pyproject.toml
[project]
name = "bullet-mania"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pygame-ce>=2.5.6",
]

[project.scripts]
bullet-mania = "bullet_mania:main"

[build-system]
requires = ["uv_build>=0.9.27,<0.10.0"]
build-backend = "uv_build"
